fix: rename Pages folder
