+++ tests/e2e/12_remove_custom_token.spec.js
@@ -1,7 +1,7 @@
 const TestUtil = require('../utils/TestUtils')
-const OverviewPage = require('../Pages/OverviewPage')
-const HomePage = require('../Pages/HomePage')
-const PasswordPage = require('../Pages/PasswordPage')
+const OverviewPage = require('../pages/OverviewPage')
+const HomePage = require('../pages/HomePage')
+const PasswordPage = require('../pages/PasswordPage')
 const expect = require('chai').expect
 
 const puppeteer = require('puppeteer')
