+++ tests/e2e/rsk_tests/06_derived_path.spec.js
@@ -1,8 +1,8 @@
 const TestUtil = require('../../utils/TestUtils')
-const OverviewPage = require('../../Pages/OverviewPage')
-const HomePage = require('../../Pages/HomePage')
-const PasswordPage = require('../../Pages/PasswordPage')
-const SeedWordsPage = require('../../Pages/SeedWordsPage')
+const OverviewPage = require('../../pages/OverviewPage')
+const HomePage = require('../../pages/HomePage')
+const PasswordPage = require('../../pages/PasswordPage')
+const SeedWordsPage = require('../../pages/SeedWordsPage')
 const expect = require('chai').expect
 
 const puppeteer = require('puppeteer')
