+++ tests/e2e/rsk_tests/13_rsk_bridge_injection.spec.js
@@ -1,7 +1,7 @@
 const TestUtil = require('../../utils/TestUtils')
-const OverviewPage = require('../../Pages/OverviewPage')
-const HomePage = require('../../Pages/HomePage')
-const PasswordPage = require('../../Pages/PasswordPage')
+const OverviewPage = require('../../pages/OverviewPage')
+const HomePage = require('../../pages/HomePage')
+const PasswordPage = require('../../pages/PasswordPage')
 const puppeteer = require('puppeteer')
 const { expect } = require('chai')
