+++ tests/e2e/terra_tests/19_terra_send.spec.js
@@ -1,10 +1,10 @@
 const TestUtil = require('../../utils/TestUtils')
-const OverviewPage = require('../../Pages/OverviewPage')
-const HomePage = require('../../Pages/HomePage')
-const PasswordPage = require('../../Pages/PasswordPage')
-const SearchAssetPage = require('../../Pages/SearchAssetPage')
-const SendPage = require('../../Pages/SendPage')
-const TransactionDetailsPage = require('../../Pages/TransactionDetailsPage')
+const OverviewPage = require('../../pages/OverviewPage')
+const HomePage = require('../../pages/HomePage')
+const PasswordPage = require('../../pages/PasswordPage')
+const SearchAssetPage = require('../../pages/SearchAssetPage')
+const SendPage = require('../../pages/SendPage')
+const TransactionDetailsPage = require('../../pages/TransactionDetailsPage')
 const puppeteer = require('puppeteer')
 
 const testUtil = new TestUtil()
