+++ tests/pages/ReceivePage.js
